pyproject.toml
[tool.poetry]
name = "madr-api"
version = "0.1.0"
description = ""
authors = ["TheCamilovisk <camilolgon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["standard"], version = "^0.115.8" }
sqlalchemy = "^2.0.38"
pydantic-settings = "^2.7.1"
alembic = "^1.14.1"
pyjwt = "^2.10.1"
tzdata = "^2025.1"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
ruff = "^0.9.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastpi dev madr_api/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=madr_api -vv'
post_test = 'coverage html'
